+++ src/config/DefaultOptions.ts
@@ -29,7 +29,6 @@ function getDefaultOptions(): IMoleculeOptions {
       fontFamily: 'Arial, Helvetica, sans-serif',
       fontSizeLarge: 11,
       fontSizeSmall: 3,
-      svgTextParity: 'pikachu',
       padding: 10.0,
       kkThreshold: 0.1,
       kkInnerThreshold: 0.1,